requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Mail==0.9.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Compress==1.14
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
passlib==1.7.4

# AI/ML
openai==1.54.0

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-asyncio==0.23.2
hypothesis==6.92.1
faker==20.1.0
selenium==4.16.0
webdriver-manager==4.0.1

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
pre-commit==3.6.0

# Security
bandit==1.7.6
safety==3.0.1
flask-limiter==3.5.0
flask-talisman==1.1.0
flask-wtf==1.2.1
pyotp==2.9.0

# Caching & Performance
redis==5.0.1
celery==5.3.4

# Monitoring
sentry-sdk[flask]==1.39.2

# Docker
docker==7.0.0

# Database Migrations
flask-migrate==4.0.5

# API Documentation
flask-swagger-ui==4.11.1
